+++ lib/legato/query/metric_filter.ex
@@ -7,10 +7,10 @@ defmodule Legato.Query.MetricFilter do
   #   "operator": enum(Operator),
   #   "comparisonValue": string,
   # }
-  defimpl Poison.Encoder, for: __MODULE__ do
+  defimpl Jason.Encoder, for: __MODULE__ do
     def encode(struct, options) do
       # This is the format for GA report json
-      Poison.Encoder.Map.encode(%{
+      Jason.Encoder.Map.encode(%{
         metric_name: struct.dimension_name,
         operator: struct.operator,
         not: struct.not,
@@ -19,5 +19,5 @@ defmodule Legato.Query.MetricFilter do
     end
   end
 
-  def to_json(filter), do: Poison.encode!(filter)
+  def to_json(filter), do: Jason.encode!(filter)
 end