+++ lib/legato/query/order.ex
@@ -1,5 +1,5 @@
 defmodule Legato.Query.Order do
-  @derive [Poison.Encoder]
+  @derive [Jason.Encoder]
   defstruct field_name: nil, order_type: :value, sort_order: :ascending
 
   # order_type: VALUE, DELTA, SMART, HISTOGRAM_BUCKET, DIMENSION_AS_INTEGER